SharpDX削除

+++ FDK19/コード/00.共通/CWin32.cs
@@ -343,21 +343,21 @@ namespace FDK
 		[DllImport( "user32.dll", CharSet = CharSet.Auto )]
 		public static extern bool IsZoomed( IntPtr hWnd );
 		[DllImport("winmm.dll")]
-		public static extern uint midiInClose(IntPtr hMidiIn);
+		public static extern uint midiInClose(uint hMidiIn);
 		[DllImport("winmm.dll")]
 		public static extern uint midiInGetDevCaps(uint uDeviceID, ref MIDIINCAPS lpMidiInCaps, uint cbMidiInCaps);
 		[DllImport("winmm.dll")]
-		public static extern uint midiInGetID(IntPtr hMidiIn, ref IntPtr puDeviceID);
+		public static extern uint midiInGetID(uint hMidiIn, ref uint puDeviceID);
 		[DllImport("winmm.dll")]
 		public static extern uint midiInGetNumDevs();
 		[DllImport("winmm.dll")]
-		public static extern uint midiInOpen(ref IntPtr phMidiIn, uint uDeviceID, MidiInProc dwCallback, IntPtr dwInstance, int fdwOpen);
+		public static extern uint midiInOpen(ref uint phMidiIn, uint uDeviceID, MidiInProc dwCallback, int dwInstance, int fdwOpen);
 		[DllImport("winmm.dll")]
-		public static extern uint midiInReset(IntPtr hMidiIn);
+		public static extern uint midiInReset(uint hMidiIn);
 		[DllImport("winmm.dll")]
-		public static extern uint midiInStart(IntPtr hMidiIn);
+		public static extern uint midiInStart(uint hMidiIn);
 		[DllImport("winmm.dll")]
-		public static extern uint midiInStop(IntPtr hMidiIn);
+		public static extern uint midiInStop(uint hMidiIn);
 		[DllImport( "user32.dll", CharSet = CharSet.Auto )]
 		public static extern IntPtr MonitorFromWindow( IntPtr hwnd, uint dwFlags );
 		[return: MarshalAs( UnmanagedType.Bool )]
@@ -656,6 +656,6 @@ namespace FDK
 
 		// Win32 メッセージ処理デリゲート
 
-		public delegate void MidiInProc(IntPtr hMidiIn, uint wMsg, IntPtr dwInstance, IntPtr dwParam1, IntPtr dwParam2);
+		public delegate void MidiInProc(uint hMidiIn, uint wMsg, int dwInstance, int dwParam1, int dwParam2);
 	}
 }
